Controllers changes
Add any controllers
Modify Dispatcher
Clean index file
Add config file
